+++ app/assets/js/.eslintrc
