+++ app/assets/stylesheets/_star-radios.scss
