+++ app/db.js
@@ -1,3 +1,3 @@
-const config = require('./knexfile.js');
+const config = require('../db/knexfile.js');
 const knex = require('knex')(config[process.env.NODE_ENV || 'development']);
 module.exports = require('bookshelf')(knex);
