+++ gulpfile.js
@@ -12,7 +12,7 @@ const uglify = require('gulp-uglify');
 let node;
 
 gulp.task('browserify', () => {
-  browserify('assets/js/main.js')
+  browserify('app/assets/js/main.js')
     .bundle()
     .on('error', function (err) {
       gutil.log(gutil.colors.red('Browserify compilation error:'));
@@ -40,7 +40,7 @@ gulp.task('fonts', () => {
 });
 
 gulp.task('sass', () => {
-  gulp.src('assets/stylesheets/*.scss')
+  gulp.src('app/assets/stylesheets/*.scss')
     .pipe(plumber())
     .pipe(
       sass({
@@ -70,9 +70,9 @@ gulp.task('server', () => {
 });
 
 gulp.task('watch', ['js', 'css', 'server'], () => {
-  gulp.watch(['routes/**/*.js', '*.js'], ['server']);
-  gulp.watch('assets/stylesheets/*.scss', ['sass']);
-  gulp.watch('assets/js/**/*.js', ['browserify']);
+  gulp.watch(['app/**/*.js', 'bin/www'], ['server']);
+  gulp.watch('app/assets/stylesheets/*.scss', ['sass']);
+  gulp.watch('app/assets/js/**/*.js', ['browserify']);
 });
