+++ package.json
@@ -9,7 +9,7 @@
     "lint": "eslint .",
     "pretest": "npm-shrinkwrap-check -3 && eslint . && npm run compile && NODE_ENV=test npm run db-migrate",
     "test": "NODE_ENV=test mocha && NODE_ENV=test npm run features",
-    "db-migrate": "knex migrate:latest",
+    "db-migrate": "knex migrate:latest --knexfile db/knexfile.js",
     "features": "NODE_ENV=test cucumber.js",
     "heroku-postbuild": "npm run compile"
   },
