+++ test/features/support/hooks.js
@@ -1,5 +1,5 @@
 const knexCleaner = require('knex-cleaner');
-const knex = require('../../../db').knex;
+const knex = require('../../../app/db').knex;
 
 module.exports = function () {
   // eslint-disable-next-line new-cap
